[State Sync] Rename synchronizer to state_synchronizer module.
